% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratedGradients.R
\name{axiom_completeness}
\alias{axiom_completeness}
\title{Verifies the axiom of completeness.}
\usage{
axiom_completeness(input, baseline, model, site, integrated.gradients)
}
\arguments{
\item{input}{The predictor field in matrix/array format.}

\item{baseline}{The integrated gradients method attributes the prediction 
at input 'x' relative to a 'baseline', computing the contribution of 'x'
to the prediction. The \code{baseline} parameter defines this baseline, . 
Default to NULL which set the baseline to a 0 array. For custom baselines, 
input an array with the dimensions matching those of the input layer of
the neural network.}

\item{model}{A keras sequential or functional model.}

\item{site}{A data frame containing the 'x' and 'y' coordinates of 
the desired site where to compute the gradients.
e.g., site = data.frame("x" = -3.82, "y" = 43.46)}

\item{integrated.gradients}{An array/matrix of integrated gradients.}
}
\value{
A matrix/array of the gradients of the predictions w.r.t input
}
\description{
Computes the difference in the prediction 
at input 'x' and the prediction at a 'baseline' and compare it
with the sum of the integrated gradients.
}
\author{
J. Bano-Medina
}
