% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernouilliGamma.loss_function.R
\name{bernouilliGamma.loss_function}
\alias{bernouilliGamma.loss_function}
\title{Custom loss function for Bernouilli-Gamma distributions.}
\usage{
bernouilliGamma.loss_function(last.connection = NULL)
}
\arguments{
\item{last.connection}{A string with values "conv" o "dense" depending on the type of the net's last connection. DEFAULT is NULL.}
}
\description{
This loss function optimizes the negative log-likelihood of a Bernouill-Gamma distribution. It is a custom
loss function defined according to keras functions.
}
\details{
This function uses \code{tensorflow} internally. Please if not, install tensorflow (>= v.)
downscaleTrain.keras for training a downscaling deep model with keras
downscalePredict.keras for predicting with a keras model
prepareNewData.keras for predictor preparation with new (test) data
\href{https://github.com/SantanderMetGroup/downscaleR.keras/wiki}{downscaleR.keras Wiki} 
\href{https://github.com/SantanderMetGroup/downscaleR/wiki/training-downscaling-models}{downscaleR Wiki} for downscaling seasonal forecasting and climate projections.
}
\seealso{
bernouilliGamma.statistics for computing the expectance or simulate from the discrete continuous distribution
}
\author{
J. Bano-Medina
}
