% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevanceMaps.R
\name{condProbDist}
\alias{condProbDist}
\title{Infer a multivariate conditional distribution.}
\usage{
condProbDist(index_k, index_l, xl, paramJoint)
}
\arguments{
\item{index_k}{Indices of the predictors of domain k.}

\item{index_l}{Indices of the predictors of domain k.}

\item{xl}{Predictors of domain l.}

\item{paramJoint}{Parameters from the joint gaussian distribution.}
}
\value{
Parameters (means and covariance matrix) from the multivariate joint gaussian distribution.
}
\description{
Infer a multivariate conditional distribution such that p(xk|xl).
}
\author{
J. Baño-Medina
}
\keyword{internal}
