% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteGammaRegular.R
\name{discreteGammaRegular}
\alias{discreteGammaRegular}
\title{Custom loss function for Bernouilli-Gamma distributions.}
\usage{
discreteGammaRegular(...)
}
\arguments{
\item{...}{An empty character}
}
\description{
This loss function optimizes the negative log-likelihood of a Bernouill-Gamma distribution. It is a custom
loss function defined according to keras functions. Use this function when working with irregular grids, otherwise 
optimize \code{\link[downscaleR.keras]{discreteGammaIrregular}}
}
\seealso{
downscaleTrain.keras for training a downscaling deep model with keras
downscalePredict.keras for predicting with a keras model
prepareNewData.keras for predictor preparation with new (test) data
\href{https://github.com/SantanderMetGroup/downscaleR/wiki/training-downscaling-models}{downscaleR Wiki} for downscaling seasonal forecasting and climate projections.

Other downscaling.functions: \code{\link{discreteGammaIrregular}},
  \code{\link{downscalePredict.keras}},
  \code{\link{downscaleTrain.keras}}
}
\author{
J. Bano-Medina
}
\concept{downscaling.functions}
