% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian.loss_function.R
\name{Gaussian.loss_function}
\alias{Gaussian.loss_function}
\title{Custom loss function for Gaussian distributions.}
\usage{
Gaussian.loss_function(last.connection = NULL)
}
\arguments{
\item{last.connection}{A string with values "conv" o "dense" depending on the type of the net's last connection. DEFAULT is NULL.}
}
\description{
This loss function optimizes the negative log-likelihood of a Gaussian distribution. It is a custom
loss function defined according to keras functions.
}
\seealso{
bernouilliGamma.loss_function for computing the negative log-likelihood of a Bernouilli-Gamma distribution
downscaleTrain.keras for training a downscaling deep model with keras
downscalePredict.keras for predicting with a keras model
prepareNewData.keras for predictor preparation with new (test) data
\href{https://github.com/SantanderMetGroup/downscaleR.keras/wiki}{downscaleR.keras Wiki} 
\href{https://github.com/SantanderMetGroup/downscaleR/wiki/training-downscaling-models}{downscaleR Wiki} for downscaling seasonal forecasting and climate projections.
}
\author{
J. Bano-Medina
}
