% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratedGradients.R
\name{get_gradients}
\alias{get_gradients}
\title{Computes the gradients of the neural network.}
\usage{
get_gradients(input, site, model)
}
\arguments{
\item{input}{The predictor field in matrix/array format.}

\item{site}{A data frame containing the 'x' and 'y' coordinates of 
the desired site where to compute the gradients.
e.g., site = data.frame("x" = -3.82, "y" = 43.46)}

\item{model}{A keras sequential or functional model.}
}
\value{
A matrix/array of the gradients of the predictions w.r.t input
}
\description{
Given a neural network, computes the gradients for a particular predictand (or 
output neuron of the model) w.r.t an input predictand field.
}
\author{
J. Bano-Medina
}
