% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.keras.R
\name{prepareData.keras}
\alias{prepareData.keras}
\title{Configuration of data for downscaling with a keras model}
\usage{
prepareData.keras(
  x,
  y,
  first.connection = c("dense", "conv"),
  last.connection = c("dense", "conv"),
  channels = c("first", "last"),
  time.frames = NULL
)
}
\arguments{
\item{x}{A grid (usually a multigrid) of predictor fields}

\item{y}{A grid (usually a stations grid, but not necessarily) of observations (predictands)}

\item{first.connection}{A string. Possible values are c("dense","conv") depending on whether 
the first connection (i.e., input layer to first hidden layer) is dense or convolutional.}

\item{last.connection}{A string. Same as \code{first.connection} but for the last connection
(i.e., last hidden layer to output layer).}

\item{channels}{A string. Possible values are c("first","last") and indicates the dimension of the channels (i.e., climate variables)
in the array. If "first" then dimensions = c("channel","latitude","longitude") for regular grids or c("channel","loc") for irregular grids.
If "last" then dimensions = c("latitude","longitude","channel") for regular grids or c("loc","channel") for irregular grids.}

\item{time.frames}{The number of time frames to build the recurrent neural network. If e.g., time.frame = 2, then the value 
y(t) is a function of x(t) and x(t-1). The time frames stack in the input array prior to the input neurons or channels (in conv. layers). 
See \code{\link[keras]{layer_simple_rnn}},\code{\link[keras]{layer_lstm}} or \code{\link[keras]{layer_conv_lstm_2d}}.}
}
\value{
A named list with components \code{y} (the predictand), \code{x.global} (global predictors) and other attributes. See Examples.
}
\description{
Configuration of data for flexible downscaling keras experiment definition
}
\details{
Remove days containing NA in at least one predictand site.
}
\examples{
\donttest{
require(climate4R.datasets) 
# Loading data
require(transformeR)
data("VALUE_Iberia_tas")
y <- VALUE_Iberia_tas
data("NCEP_Iberia_hus850", "NCEP_Iberia_psl", "NCEP_Iberia_ta850")
x <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_psl, NCEP_Iberia_ta850)
# We standardize the predictors using transformeR function scaleGrid
x <- scaleGrid(x,type = "standardize")
# Preparing the predictors
data <- prepareData.keras(x = x, y = y, 
                          first.connection = "conv",
                          last.connection = "dense",
                          channels = "last")
# We can visualize the outputield not imported f
str(data)
}
}
\seealso{
downscaleTrain.keras for training a downscaling deep model with keras
downscalePredict.keras for predicting with a keras model
prepareNewData.keras for predictor preparation with new (test) data
\href{https://github.com/SantanderMetGroup/downscaleR.keras/wiki}{downscaleR.keras Wiki}

\href{https://github.com/SantanderMetGroup/downscaleR/wiki/preparing-predictor-data}{downscaleR Wiki} for preparing predictors for downscaling and seasonal forecasting.
}
\author{
J. Baño-Medina
}
\concept{downscaling.helpers}
