% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevanceMaps.R
\name{relevanceMaps}
\alias{relevanceMaps}
\title{Obtain relevance maps as a climate4R object}
\usage{
relevanceMaps(x, obj, C4R.template, model, bernouilliGamma = FALSE,
  parch = c("channel", "variable", "all"), l, num_samples)
}
\arguments{
\item{x}{A grid (usually a multigrid) of predictor fields.}

\item{obj}{The object as returned by \code{\link[downscaleR.keras]{prepareData.keras}}.}

\item{C4R.template}{A climate4R grid that serves as template for the returned prediction object.}

\item{model}{A keras sequential or functional model.}

\item{bernouilliGamma}{A logical value. Indicates whether the \code{\link[downscaleR.keras]{bernouilliGamma.loss_function}}
was used to train the model in \code{\link[downscaleR.keras]{downscaleTrain.keras}}. Default is FALSE.}

\item{parch}{Possible values are c("all","variable","channel"). Indicates whether we want to marginalize the influence of
a certain gridpoint as a whole ("all"), to isolate the influence per variable 
("variable", i.e., for example to isolate the influence of the specific humidity at all levels: hus500,hus700,...) 
or of every channel (i.e., "channel") independently.}

\item{l}{A numeric value. Defines the domain ((2l+1)x(2l+1)) used to infer the conditional multivariate gaussian distribution}

\item{num_samples}{A numeric value. How many times do we sample from the multivariate gaussian distribution.}
}
\value{
A climate 4R object containing the relevance maps as members. The attribute attr(out,"memberCoords") is added to the climate4R output object
containing the coordinates in sequential order, such as the relevance map of the first member measures the influence
on the first coordinate in attr(out,"memberCoords").
\href{https://github.com/SantanderMetGroup/downscaleR.keras/wiki}{downscaleR.keras Wiki}
}
\description{
Relevance maps are a representation of the influence of the predictor variables on the downscaling over 
a particular gridpoint on a certain day. The relevance maps are obtained using 
\href{https://arxiv.org/pdf/1702.04595.pdf}{prediction difference analysis}
on a trained deep model (see \code{\link[downscaleR.keras]{downscaleTrain.keras}}).
}
\details{
This function relies on keras, which is a high-level neural networks API capable of running on top of tensorflow, CNTK or theano.
There are official \href{https://keras.rstudio.com/}{keras tutorials} regarding how to build deep learning models. We suggest the user, especially the beginners,
to consult these tutorials before using the downscaleR.keras package. Moreover, we encourage the reader to consult 
the prediction difference analysis technique employed 
which is published in this \href{https://arxiv.org/pdf/1702.04595.pdf}{paper}.
}
\seealso{
Other downscaling.functions: \code{\link{downscalePredict.keras}},
  \code{\link{downscaleTrain.keras}}
}
\author{
J. Bano-Medina
}
\concept{downscaling.functions}
