% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_spatial_concrete_dropout.R
\name{spatialConcreteDropout}
\alias{spatialConcreteDropout}
\title{Custom class layer with spatial concrete dropout.}
\description{
This is a keras layer enabling to implement concrete dropout 
on the convolutinal layers of the deep learning model.
}
\details{
This is an adaptation of the python spatial concrete dropout class
to the R software, developed in 
https://github.com/yaringal/ConcreteDropout/blob/master/spatial-concrete-dropout-keras.ipynb
}
\author{
J. Bano-Medina
}
\section{Super class}{
\code{\link[keras:KerasWrapper]{keras::KerasWrapper}} -> \code{spatialConcreteDropout}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{spatialConcreteDropout$new()}}
\item \href{#method-build}{\code{spatialConcreteDropout$build()}}
\item \href{#method-spatial_concrete_dropout}{\code{spatialConcreteDropout$spatial_concrete_dropout()}}
\item \href{#method-call}{\code{spatialConcreteDropout$call()}}
\item \href{#method-clone}{\code{spatialConcreteDropout$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="keras" data-topic="KerasWrapper" data-id=".set_py_wrapper">}\href{../../keras/html/KerasWrapper.html#method-.set_py_wrapper}{\code{keras::KerasWrapper$.set_py_wrapper()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="keras" data-topic="KerasWrapper" data-id="add_loss">}\href{../../keras/html/KerasWrapper.html#method-add_loss}{\code{keras::KerasWrapper$add_loss()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="keras" data-topic="KerasWrapper" data-id="add_weight">}\href{../../keras/html/KerasWrapper.html#method-add_weight}{\code{keras::KerasWrapper$add_weight()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="keras" data-topic="KerasWrapper" data-id="compute_output_shape">}\href{../../keras/html/KerasWrapper.html#method-compute_output_shape}{\code{keras::KerasWrapper$compute_output_shape()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="keras" data-topic="KerasWrapper" data-id="python_wrapper">}\href{../../keras/html/KerasWrapper.html#method-python_wrapper}{\code{keras::KerasWrapper$python_wrapper()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialConcreteDropout$new(
  weight_regularizer,
  dropout_regularizer,
  init_min,
  init_max,
  is_mc_dropout,
  data.format
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build"></a>}}
\if{latex}{\out{\hypertarget{method-build}{}}}
\subsection{Method \code{build()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialConcreteDropout$build(input_shape)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spatial_concrete_dropout"></a>}}
\if{latex}{\out{\hypertarget{method-spatial_concrete_dropout}{}}}
\subsection{Method \code{spatial_concrete_dropout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialConcreteDropout$spatial_concrete_dropout(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-call"></a>}}
\if{latex}{\out{\hypertarget{method-call}{}}}
\subsection{Method \code{call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialConcreteDropout$call(x, mask = NULL, training = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{spatialConcreteDropout$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
