% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevanceMaps.R
\name{condProbDist}
\alias{condProbDist}
\title{Infer a multivariate conditional distribution.}
\usage{
condProbDist(index_k, index_l, xl, paramJoint)
}
\arguments{
\item{xl}{Predictors of a domain of size ((2l+1)x(2l+1))}

\item{xk}{Predictors of a domain of size ((2k+1)x(2k+1))}

\item{num_samples}{A numeric value. How many times do we sample from the multivariate gaussian distribution.}
}
\value{
Parameters (means and covariance matrix) from the multivariate joint gaussian distribution.
}
\description{
Infer a multivariate conditional distribution such that p(xk|xl).
}
\author{
J. Baño-Medina
}
\keyword{internal}
