% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevanceMaps.R
\name{jointProbDist}
\alias{jointProbDist}
\title{Infer a multivariate joint distribution.}
\usage{
jointProbDist(xl)
}
\arguments{
\item{xl}{Predictors of domain l.}
}
\value{
Parameters (means and covariance matrix) from the multivariate joint gaussian distribution.
}
\description{
Infer a multivariate joint distribution such that p(xl).
}
\author{
J. Baño-Medina
}
\keyword{internal}
