% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialConcreteDropout.R
\name{layer_spatial_concrete_dropout}
\alias{layer_spatial_concrete_dropout}
\title{Concrete dropout for convolutional keras layers.}
\usage{
layer_spatial_concrete_dropout(
  object,
  layer,
  weight_regularizer = 1e-06,
  dropout_regularizer = 1e-05,
  init_min = 0.1,
  init_max = 0.1,
  is_mc_dropout = TRUE,
  name = NULL,
  trainable = TRUE,
  data.format = "channels_last"
)
}
\arguments{
\item{object}{Model or layer object}

\item{layer}{A layer instance.}

\item{weight_regularizer}{A positive number that penalizes the weights
of the layer according to L1 or L2 regularization.}

\item{dropout_regularizer}{A positive number adequate to initialize the dropout probability}

\item{init_min}{.}

\item{init_max}{.}

\item{is_mc_dropout}{Wether Monte-Carlo dropout is applied or not. Default to TRUE.}

\item{name}{An optional name string for the layer. 
Should be unique in a model (do not reuse the same name twice). 
It will be autogenerated if it isn't provided.}

\item{trainable}{Whether the layer weights will be updated during training.}

\item{data.format}{A string, one of channels_last (default) or channels_first. 
The ordering of the dimensions in the inputs. channels_last corresponds to inputs with 
shape (batch, height, width, channels) while channels_first corresponds to inputs with shape (batch, channels, height, width). 
It defaults to the image_data_format value found in your Keras config file at ~/.keras/keras.json. 
If you never set it, then it will be "channels_last".}
}
\description{
This function is for instantiating custom spatial concrete dropout layer. 
It allows to learn the dropout probability for a given `Convolutional 2D' layer.
}
\details{
This is an adaptation of the python spatial concrete dropout class
to the R software, developed in 
https://github.com/yaringal/ConcreteDropout/blob/master/spatial-concrete-dropout-keras.ipynb
}
\author{
J. Bano-Medina
}
